Fix unsupported frame type, Refactor main to c++
I fixed the "wifi:unsupported frame type c0" by refactoring the deauth
program for c++ and adding an extern definition to overwrite
ieee80211_raw_frame_sanity_check(). Initially, this failed to compile
because of a multiple definitions error so I added a target linker to
the CMakeLists.txt file to allow for multiple definitions.