Add oled ui prototype
I've added a prototype for my ui that allows me to display options on
the screen and select them with buttons. I did this by using the u8g2
library for the screen and buttons for user input. I made structs for
both pages and textlines to make it easier to load pages. This program
uses the freertos isr handler to take interrupts from each push button.
When a push button is pressed, the gpio of that push button is added to
a queue which is received by the ButtonTask which updates the display as
needed with the display_page function.